WIP configure sync method
